Rename `OverallTable` to `OverallSummary`
