+++ site/frontend/src/pages/compare/summary/overall-summary.vue
